refactor: 잔돈반환 섹션 컴포넌트 분리
